fix: add http-errors dependency, move recaptcha tests into __tests__ directory, use deepmerge in renderer.tsx
