+++ package.json
@@ -131,6 +131,7 @@
     "express": "^4.17.1",
     "express-recaptcha": "^5.0.2",
     "helmet": "^4.6.0",
+    "http-errors": "^1.8.1",
     "js-sha1": "^0.6.0",
     "jsonwebtoken": "^8.5.1",
     "mjml": "^4.10.4",
